pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "orewatch"
version = "1.1.0"
description = "OreWatch: multi-ecosystem malicious package detection and background monitoring for developers."
readme = "README.md"
requires-python = ">=3.14"
license = "MIT"
license-files = ["LICENSE"]
authors = [
  { name = "rapticore" }
]
keywords = [
  "security",
  "supply-chain",
  "malware",
  "dependencies",
  "scanner",
  "sbom",
  "monitoring"
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "Operating System :: MacOS",
  "Operating System :: POSIX :: Linux",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.14",
  "Topic :: Security",
  "Topic :: Software Development :: Quality Assurance",
  "Environment :: Console",
]
dependencies = [
  "PyYAML>=6.0,<7.0",
  "requests>=2.31.0",
  "beautifulsoup4>=4.12.0",
  "packaging>=21.0",
]

[project.urls]
Homepage = "https://github.com/rapticore/ore-mal-pkg-inspector"
Repository = "https://github.com/rapticore/ore-mal-pkg-inspector"
Issues = "https://github.com/rapticore/ore-mal-pkg-inspector/issues"

[project.scripts]
orewatch = "malicious_package_scanner:main"
ore-mal-pkg-inspector = "malicious_package_scanner:main"

[tool.setuptools]
py-modules = ["logging_config", "malicious_package_scanner", "scanner_engine"]

[tool.setuptools.packages.find]
include = ["collectors", "monitor", "scanners"]

[tool.setuptools.package-data]
collectors = ["config.yaml"]
